requirements.txt
Flask
Flask-Bootstrap
flask-WTF
pymongo
seaborn
matplotlib
wordcloud
numpy
tweepy
pandas
pillow
mpld3
nltk
